perf: update docs and npm update support
This commit adds a new README to the project that better defines the
scope and intent of what AuroKit is and how it will support consuming
developers.

Additionally there is an update to the ./package.json file that replaces
the `^` with `~` to restrict `npm update` to only PATCH releases.

On branch dsande/generalImprovements
Changes to be committed:
modified:   README.md
new file:   package-lock.json
modified:   package.json